Updated test file for Mac/Windows testing
